Move all the "secure.mk" into Makefile, update docs
There's no reason to keep it all hidden since that was for legacy
reasons. There's nothing fancy going on in there, and never was, so
let's just move all into one place.

Along the way the `clean-plain` target was changed to accomodate the
above and to also delete the `decrypt_pw`. This means that the
[README.md](./README.md), which needed to be updated for the secure.mk
changes, could remove the `rm -f decrypt_pw` notes.

Signed-off-by: Lisa Seelye <lisa@users.noreply.github.com>